// src/types/cart.ts
export interface CurrencyInfo {
  currency_code: CurrencyCode;
  currency_symbol: string;
  currency_minor_unit: number;
  currency_decimal_separator: string;
  currency_thousand_separator: string;
  currency_prefix: string;
  currency_suffix: string;
}
// ...
export interface CartImageItem {
  id: number;
  src: string;
  thumbnail: string;
  srcset: string;
  sizes: string;
  name: string;
  alt: string;
}

export interface CartVariationItem {
  attribute: string;
  value: string;
  raw_attribute: string;
}

export interface CartItemPrices extends CurrencyInfo {
  price: string;
  regular_price: string;
  sale_price: string;
  price_range: null | { min_amount: string; max_amount: string };
  raw_prices: {
    precision: number;
    price: string;
    regular_price: string;
    sale_price: string;
  };
}

export interface CartItemTotals extends CurrencyInfo {
  line_subtotal: string;
  line_subtotal_tax: string;
  line_total: string;
  line_total_tax: string;
}

export type CatalogVisibility = 'catalog' | 'hidden' | 'search' | 'visible';

export interface CartItem {
  key: string;
  id: number;
  type: string;
  quantity: number;
  catalog_visibility: CatalogVisibility;
  quantity_limits: {
    minimum: number;
    maximum: number;
    multiple_of: number;
    editable: boolean;
  };
  name: string;
  summary: string;
  short_description: string;
  description: string;
  sku: string;
  low_stock_remaining: null | number;
  backorders_allowed: boolean;
  show_backorder_badge: boolean;
  sold_individually: boolean;
  permalink: string;
  images: Array<CartImageItem>;
  variation: Array<CartVariationItem>;
  prices: CartItemPrices;
  totals: CartItemTotals;
  extensions: ExtensionsData;
  item_data: ProductResponseItemData[];
}